old terrain style mostly implemented minus shadows

+++ shaders/dx11/vertex/terrain_vs.bin.hlsl
@@ -612,7 +612,7 @@ float2 texture2DBc5(BgfxSampler2D _sampler, float2 _uv)
 {
 return bgfxTexture2D(_sampler, _uv).xy;
 }
-Output main( float3 a_normal : NORMAL , float3 a_position : POSITION , float2 a_texcoord0 : TEXCOORD0 , int a_texcoord2 : TEXCOORD2 , int a_texcoord3 : TEXCOORD3 , int4 a_texcoord4 : TEXCOORD4) { Output _varying_; _varying_.v_adjacentMatIndices; _varying_.v_blendMode; _varying_.v_materialID; _varying_.v_normal; _varying_.v_texcoord0; _varying_.v_worldPos;
+Output main( float3 a_normal : NORMAL , float3 a_position : POSITION , float2 a_texcoord0 : TEXCOORD0 , float a_texcoord2 : TEXCOORD2 , float a_texcoord3 : TEXCOORD3 , int4 a_texcoord4 : TEXCOORD4) { Output _varying_; _varying_.v_adjacentMatIndices; _varying_.v_blendMode; _varying_.v_materialID; _varying_.v_normal; _varying_.v_texcoord0; _varying_.v_worldPos;
 {
 _varying_.v_texcoord0 = a_texcoord0;
 _varying_.v_worldPos = mul(u_model[0], float4(a_position, 1.0) ).xyz;
